+++ django/applications/catmaid/static/js/skeleton-model.js
@@ -19,6 +19,7 @@
       this.pre_visible = true;
       this.post_visible = true;
       this.desmosome_visible = true;
+      this.mitochondrion_visible = true;
       this.text_visible = false;
       this.meta_visible = true;
       this.color = color || new THREE.Color(defaultColor);
@@ -42,6 +43,7 @@
       this.pre_visible = v;
       this.post_visible = v;
       this.desmosome_visible = v;
+      this.mitochondrion_visible = v;
       if (!v) this.text_visible = v;
       this.meta_visible = v;
   };
@@ -52,6 +54,7 @@
     m.pre_visible = this.pre_visible;
     m.post_visible = this.post_visible;
     m.desmosome_visible = this.desmosome_visible;
+    m.mitochondrion_visible = this.mitochondrion_visible;
     m.text_visible = this.text_visible;
     m.meta_visible = this.meta_visible;
     m.opacity = this.opacity;
@@ -75,6 +78,7 @@
     this.pre_visible = other.pre_visible;
     this.post_visible = other.post_visible;
     this.desmosome_visible = other.desmosome_visible;
+    this.mitochondrion_visible = other.mitochondrion_visible;
     this.text_visible = other.text_visible;
     this.meta_visible = other.meta_visible;
     this.opacity = other.opacity;